Renaming the package to bayte
