+++ .github/workflows/test-package.yml
@@ -30,11 +30,11 @@ jobs:
             - name: Install dependencies
               run: python -m pip install -r requirements.txt .[dev]
             - name: Check docstrings
-              run: python -m pydocstyle bayesian_target --convention=numpy
+              run: python -m pydocstyle bayte --convention=numpy
             - name: Check static typing
-              run: python -m mypy bayesian_target
+              run: python -m mypy bayte
             - name: Run flake8
-              run: python -m flake8 bayesian_target
+              run: python -m flake8 bayte
             - name: Run unit testing
               run: python -m pytest tests --cov=./ --cov-report=xml
             - name: Run Codecov
