--- bayesian_target/__init__.py
@@ -1,10 +1,0 @@
-"""Import path."""
-
-from ._meta import __version__  # noqa: F401
-
-from typing import List
-
-from .encoder import BayesianTargetEncoder
-from .ensemble import BayesianTargetEstimator
-
-__all__: List[str] = ["BayesianTargetEncoder", "BayesianTargetEstimator"]
