+++ bayte/_meta.py
