+++ setup.cfg
@@ -1,12 +1,12 @@
 [metadata]
-name = bayesian-target
-version = attr: bayesian_target._meta.__version__
+name = bayte
+version = attr: bayte._meta.__version__
 description = Bayesian target encoding with scikit-learn and scipy
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = Akshay Gupta
 author_email = akgcodes@gmail.com
-url = https://github.com/ak-gupta/bayesian-target
+url = https://github.com/ak-gupta/bayte
 python_requires = 
     >=3.7
 license = MIT license
@@ -59,7 +59,7 @@ version_pattern = "MAJOR.MINOR.PATCH"
 [bumpver:file_patterns]
 setup.cfg = 
     current_version = "{version}"
-bayesian_target/_meta.py = 
+bayte/_meta.py = 
     __version__ = "{version}"
 
 [flake8]
